pyproject.toml
[project]
name = "research-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Nguyen Anh",email = "tuananh835.nta@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "opensearch-py (>=3.0.0,<4.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langgraph (>=0.6.8,<0.7.0)",
    "pydantic-ai (>=1.0.13,<2.0.0)",
    "docling (>=2.55.0,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "ollama (>=0.6.0,<0.7.0)",
    "litellm (>=1.77.7,<2.0.0)",
    "groq (>=0.32.0,<0.33.0)"
]

[tool.poetry]
packages = [{include = "research_assistant", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
